Randomly choose between current lane and adjacent.
This doesn't work very well because we're (1) randomly choosing the adjacent lane, and (2) randomly choosing whether to do the adjacent-lane plan or cruise in the current lane. So, the decision (2) doesn't have much persistence and is usually overridden in the next planning session. This is fine though, since presumably we'll have a better scheme for (2) soon.